refactor(social): remove Follow Trading feed feature flag
Always render Leaderboard | Feed tabs now that the feed is fully rolled out, and drop the unused standalone leaderboard shell.
