+++ app/components/Nav/Main/MainNavigator.test.tsx
@@ -1238,7 +1238,7 @@ describe('MainNavigator', () => {
     });
   });
 
-  it('includes SocialTradersView screen when Social Leaderboard remote flag is enabled', () => {
+  it('includes SocialTradersTabsView screen when Social Leaderboard remote flag is enabled', () => {
     const stateWithSocialLeaderboard = {
       ...initialRootState,
       engine: {
@@ -1287,7 +1287,7 @@ describe('MainNavigator', () => {
     );
 
     expect(topTradersScreen).toBeDefined();
-    expect(topTradersScreen?.component.name).toBe('SocialTradersView');
+    expect(topTradersScreen?.component.name).toBe('SocialTradersTabsView');
   });
 
   describe('Inner navigator component rendering', () => {
