+++ app/components/Views/SocialLeaderboard/FeedView/FeedView.tsx
@@ -142,8 +142,8 @@ const FeedView: React.FC<FeedViewProps> = ({
   const navigation = useNavigation<AppNavigationProp>();
   // `'TopTradersView'` is the *route* name for the whole Follow Trading surface
   // (`Routes.SOCIAL_LEADERBOARD.VIEW`), not the sibling component of the same
-  // name — the feed renders inside it via `SocialTradersView`, so this is the
-  // enclosing route's param list even though the names look mismatched.
+  // name — the feed renders inside it via `SocialTradersTabsView`, so this is
+  // the enclosing route's param list even though the names look mismatched.
   const route = useRoute<RouteProp<RootStackParamList, 'TopTradersView'>>();
   const { track } = useSocialLeaderboardAnalytics();
   const source = route.params?.source ?? 'nav_tab';
