+++ app/components/Views/SocialLeaderboard/SocialTradersTabsView/SocialTradersTabsView.test.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import { act, fireEvent, screen } from '@testing-library/react-native';
 import renderWithProvider from '../../../../util/test/renderWithProvider';
 import { MetaMetricsEvents } from '../../../../core/Analytics';
+import Routes from '../../../../constants/navigation/Routes';
 import SocialTradersTabsView from './SocialTradersTabsView';
 import { SocialTradersTabsViewSelectorsIDs } from './SocialTradersTabsView.testIds';
 import { SCROLLABLE_SCREEN_SAFE_AREA_EDGES } from '../shared/scrollableScreenSafeArea';
@@ -9,6 +10,11 @@ import { expectHeaderIncludesTopInset } from '../shared/scrollableScreenSafeArea
 
 const mockPlaySelection = jest.fn().mockResolvedValue(undefined);
 const mockTrack = jest.fn();
+const mockGoBack = jest.fn();
+const mockNavigate = jest.fn();
+const mockOpenSystemSettings = jest.fn();
+const mockHasNotificationPreferences = jest.fn(() => false);
+let mockRouteParams: { showNotificationsBanner?: boolean } = {};
 
 jest.mock('../analytics', () => {
   const actual = jest.requireActual('../analytics');
@@ -43,6 +49,29 @@ jest.mock('../../../../util/haptics', () => ({
   playSelection: () => mockPlaySelection(),
 }));
 
+jest.mock(
+  '../../../../util/notifications/services/NotificationService',
+  () => ({
+    __esModule: true,
+    default: { openSystemSettings: () => mockOpenSystemSettings() },
+  }),
+);
+
+jest.mock('@react-navigation/native', () => {
+  const actual = jest.requireActual('@react-navigation/native');
+  let navigation: { goBack: jest.Mock; navigate: jest.Mock } | undefined;
+  return {
+    ...actual,
+    useNavigation: () => {
+      if (!navigation) {
+        navigation = { goBack: mockGoBack, navigate: mockNavigate };
+      }
+      return navigation;
+    },
+    useRoute: () => ({ params: mockRouteParams, name: 'TopTradersView' }),
+  };
+});
+
 jest.mock('../../../../../locales/i18n', () => ({
   strings: (key: string) => key,
 }));
@@ -51,7 +80,7 @@ jest.mock(
   '../../Settings/NotificationsSettings/hooks/useNotificationStoragePreferences',
   () => ({
     useNotificationStoragePreferences: () => ({
-      hasNotificationPreferences: false,
+      hasNotificationPreferences: mockHasNotificationPreferences(),
       isLoading: false,
     }),
   }),
@@ -146,6 +175,8 @@ describe('SocialTradersTabsView', () => {
     mockDeferBuyActionRef = false;
     mockAttachBuyActionRef = null;
     mockOnSpotAvailabilityChange = undefined;
+    mockHasNotificationPreferences.mockReturnValue(false);
+    mockRouteParams = {};
   });
 
   it('renders the header, tabs, and both pages', () => {
@@ -189,6 +220,129 @@ describe('SocialTradersTabsView', () => {
     expect(
       screen.getByTestId(SocialTradersTabsViewSelectorsIDs.TITLE),
     ).toHaveTextContent('social_leaderboard.feed.title');
+  });
+
+  it('calls goBack when the back button is pressed', () => {
+    renderWithProvider(<SocialTradersTabsView />);
+
+    fireEvent.press(
+      screen.getByTestId(SocialTradersTabsViewSelectorsIDs.BACK_BUTTON),
+    );
+
+    expect(mockGoBack).toHaveBeenCalledTimes(1);
+  });
+
+  it('navigates to the socialAI notification settings section when the bell is pressed and preferences exist', () => {
+    mockHasNotificationPreferences.mockReturnValue(true);
+
+    renderWithProvider(<SocialTradersTabsView />);
+    fireEvent.press(
+      screen.getByTestId(SocialTradersTabsViewSelectorsIDs.NOTIFICATION_BUTTON),
+    );
+
+    expect(mockNavigate).toHaveBeenCalledWith(Routes.SETTINGS_VIEW, {
+      screen: Routes.SETTINGS.NOTIFICATION_SETTINGS_SECTION,
+      params: {
+        type: 'socialAI',
+        title: 'app_settings.notifications_opts.social_ai_title',
+        description: 'app_settings.notifications_opts.social_ai_desc',
+      },
+    });
+  });
+
+  it('navigates to notification settings when the bell is pressed and preferences do not exist yet', () => {
+    mockHasNotificationPreferences.mockReturnValue(false);
+
+    renderWithProvider(<SocialTradersTabsView />);
+    fireEvent.press(
+      screen.getByTestId(SocialTradersTabsViewSelectorsIDs.NOTIFICATION_BUTTON),
+    );
+
+    expect(mockNavigate).toHaveBeenCalledWith(Routes.SETTINGS_VIEW, {
+      screen: Routes.SETTINGS.NOTIFICATIONS,
+    });
+  });
+
+  describe('notifications nudge banner', () => {
+    it('is hidden by default when the route param is unset', () => {
+      renderWithProvider(<SocialTradersTabsView />);
+
+      expect(
+        screen.queryByTestId(
+          SocialTradersTabsViewSelectorsIDs.NOTIFICATIONS_BANNER,
+        ),
+      ).toBeNull();
+    });
+
+    it('renders when the showNotificationsBanner route param is set', () => {
+      mockRouteParams = { showNotificationsBanner: true };
+
+      renderWithProvider(<SocialTradersTabsView />);
+
+      expect(
+        screen.getByTestId(
+          SocialTradersTabsViewSelectorsIDs.NOTIFICATIONS_BANNER,
+        ),
+      ).toBeOnTheScreen();
+    });
+
+    it('opens system settings and dismisses when the CTA is pressed', () => {
+      mockRouteParams = { showNotificationsBanner: true };
+
+      renderWithProvider(<SocialTradersTabsView />);
+      fireEvent.press(
+        screen.getByText(
+          'social_leaderboard.top_traders_view.notifications_banner.open_settings',
+        ),
+      );
+
+      expect(mockOpenSystemSettings).toHaveBeenCalledTimes(1);
+      expect(
+        screen.queryByTestId(
+          SocialTradersTabsViewSelectorsIDs.NOTIFICATIONS_BANNER,
+        ),
+      ).toBeNull();
+    });
+
+    it('dismisses when the close button is pressed', () => {
+      mockRouteParams = { showNotificationsBanner: true };
+
+      renderWithProvider(<SocialTradersTabsView />);
+      fireEvent.press(screen.getByLabelText('Close banner'));
+
+      expect(mockOpenSystemSettings).not.toHaveBeenCalled();
+      expect(
+        screen.queryByTestId(
+          SocialTradersTabsViewSelectorsIDs.NOTIFICATIONS_BANNER,
+        ),
+      ).toBeNull();
+    });
+
+    it('auto-dismisses after the timeout window', () => {
+      jest.useFakeTimers();
+      try {
+        mockRouteParams = { showNotificationsBanner: true };
+        renderWithProvider(<SocialTradersTabsView />);
+
+        expect(
+          screen.getByTestId(
+            SocialTradersTabsViewSelectorsIDs.NOTIFICATIONS_BANNER,
+          ),
+        ).toBeOnTheScreen();
+
+        act(() => {
+          jest.advanceTimersByTime(20000);
+        });
+
+        expect(
+          screen.queryByTestId(
+            SocialTradersTabsViewSelectorsIDs.NOTIFICATIONS_BANNER,
+          ),
+        ).toBeNull();
+      } finally {
+        jest.useRealTimers();
+      }
+    });
   });
 
   it('plays a selection haptic when switching to a different tab', () => {
