+++ app/components/Views/SocialLeaderboard/TopTradersView/TopTradersView.test.tsx
@@ -14,8 +14,6 @@ import { MetaMetricsEvents } from '../../../../core/Analytics';
 import { ImpactMoment } from '../../../../util/haptics';
 import TopTradersView from './TopTradersView';
 import { TopTradersViewSelectorsIDs } from './TopTradersView.testIds';
-import { SCROLLABLE_SCREEN_SAFE_AREA_EDGES } from '../shared/scrollableScreenSafeArea';
-import { expectHeaderIncludesTopInset } from '../shared/scrollableScreenSafeArea.testUtils';
 import {
   getSortFilterOptionTestId,
   getTimeframeFilterOptionTestId,
@@ -103,19 +101,9 @@ const mockNavigate = jest.fn();
 const mockToggleFollow = jest.fn();
 const mockRefresh = jest.fn();
 const mockHasNotificationPreferences = jest.fn(() => true);
-const mockOpenSystemSettings = jest.fn();
 let mockRouteParams: {
   source?: string;
-  showNotificationsBanner?: boolean;
 } = {};
-
-jest.mock(
-  '../../../../util/notifications/services/NotificationService',
-  () => ({
-    __esModule: true,
-    default: { openSystemSettings: () => mockOpenSystemSettings() },
-  }),
-);
 
 const mockToggleTraderNotification = jest.fn();
 const mockIsTraderNotificationEnabled = jest.fn((_traderId: string) => true);
@@ -338,190 +326,6 @@ describe('TopTradersView', () => {
     mockRouteParams = {};
     mockNotificationPreferences = { ...defaultNotificationPreferences };
     mockIsTraderNotificationEnabled.mockReturnValue(true);
-  });
-
-  it('renders the container', () => {
-    renderWithProvider(<TopTradersView />);
-    expect(
-      screen.getByTestId(TopTradersViewSelectorsIDs.CONTAINER),
-    ).toBeOnTheScreen();
-  });
-
-  describe('safe area layout', () => {
-    it('excludes bottom safe area so the scroll list extends to the screen edge', () => {
-      renderWithProvider(<TopTradersView />);
-
-      expect(
-        screen.getByTestId(TopTradersViewSelectorsIDs.CONTAINER).props.edges,
-      ).toEqual(SCROLLABLE_SCREEN_SAFE_AREA_EDGES);
-    });
-
-    it('keeps the top inset on the header to prevent layout shift on push', () => {
-      renderWithProvider(<TopTradersView />);
-
-      expectHeaderIncludesTopInset(
-        screen.getByTestId(TopTradersViewSelectorsIDs.HEADER),
-      );
-    });
-  });
-
-  it('renders the Weekly top traders title in the scrollable title section', () => {
-    renderWithProvider(<TopTradersView />);
-
-    expect(
-      screen.getByTestId(TopTradersViewSelectorsIDs.TITLE),
-    ).toHaveTextContent('Weekly top traders');
-  });
-
-  it('connects the scrollable title section to the compact header', () => {
-    renderWithProvider(<TopTradersView />);
-
-    act(() => {
-      fireEvent(
-        screen.getByTestId(TopTradersViewSelectorsIDs.TITLE_SECTION_WRAPPER),
-        'layout',
-        {
-          nativeEvent: { layout: { height: 64 } },
-        },
-      );
-    });
-
-    expect(
-      screen.getByTestId(TopTradersViewSelectorsIDs.HEADER_TITLE),
-    ).toHaveTextContent('Weekly top traders');
-    expect(
-      screen.getByTestId(TopTradersViewSelectorsIDs.TRADER_LIST).props.onScroll,
-    ).toEqual(expect.any(Function));
-    expect(
-      screen.getByTestId(TopTradersViewSelectorsIDs.TRADER_LIST).props
-        .scrollEventThrottle,
-    ).toBe(16);
-  });
-
-  it('renders a pinned filter bar with its own type selector test ID', () => {
-    renderWithProvider(<TopTradersView />);
-
-    expect(
-      screen.getByTestId(TopTradersViewSelectorsIDs.PINNED_FILTER_BAR, {
-        includeHiddenElements: true,
-      }),
-    ).toBeOnTheScreen();
-    // The header selector and the pinned selector use distinct test IDs, so the
-    // primary selector resolves to exactly one node.
-    expect(
-      screen.getAllByTestId(TopTradersViewSelectorsIDs.TYPE_SELECTOR),
-    ).toHaveLength(1);
-  });
-
-  it('calls goBack when the back button is pressed', () => {
-    renderWithProvider(<TopTradersView />);
-    fireEvent.press(screen.getByTestId(TopTradersViewSelectorsIDs.BACK_BUTTON));
-    expect(mockGoBack).toHaveBeenCalledTimes(1);
-  });
-
-  it('renders the notification button', () => {
-    renderWithProvider(<TopTradersView />);
-    expect(
-      screen.getByTestId(TopTradersViewSelectorsIDs.NOTIFICATION_BUTTON),
-    ).toBeOnTheScreen();
-  });
-
-  describe('notifications nudge banner', () => {
-    it('is hidden by default (no route param)', () => {
-      renderWithProvider(<TopTradersView />);
-      expect(
-        screen.queryByTestId(TopTradersViewSelectorsIDs.NOTIFICATIONS_BANNER),
-      ).toBeNull();
-    });
-
-    it('renders when the showNotificationsBanner route param is set', () => {
-      mockRouteParams = { showNotificationsBanner: true };
-      renderWithProvider(<TopTradersView />);
-
-      expect(
-        screen.getByTestId(TopTradersViewSelectorsIDs.NOTIFICATIONS_BANNER),
-      ).toBeOnTheScreen();
-      expect(
-        screen.getByText(
-          "You'll get alerts when traders you follow make a move. Turn on notifications in your device settings.",
-        ),
-      ).toBeOnTheScreen();
-    });
-
-    it('opens system settings and dismisses when the CTA is pressed', () => {
-      mockRouteParams = { showNotificationsBanner: true };
-      renderWithProvider(<TopTradersView />);
-
-      fireEvent.press(screen.getByText('Open settings'));
-
-      expect(mockOpenSystemSettings).toHaveBeenCalledTimes(1);
-      expect(
-        screen.queryByTestId(TopTradersViewSelectorsIDs.NOTIFICATIONS_BANNER),
-      ).toBeNull();
-    });
-
-    it('dismisses when the close button is pressed', () => {
-      mockRouteParams = { showNotificationsBanner: true };
-      renderWithProvider(<TopTradersView />);
-
-      fireEvent.press(screen.getByLabelText('Close banner'));
-
-      expect(mockOpenSystemSettings).not.toHaveBeenCalled();
-      expect(
-        screen.queryByTestId(TopTradersViewSelectorsIDs.NOTIFICATIONS_BANNER),
-      ).toBeNull();
-    });
-
-    it('auto-dismisses after the timeout window', () => {
-      jest.useFakeTimers();
-      try {
-        mockRouteParams = { showNotificationsBanner: true };
-        renderWithProvider(<TopTradersView />);
-
-        expect(
-          screen.getByTestId(TopTradersViewSelectorsIDs.NOTIFICATIONS_BANNER),
-        ).toBeOnTheScreen();
-
-        act(() => {
-          jest.advanceTimersByTime(20000);
-        });
-
-        expect(
-          screen.queryByTestId(TopTradersViewSelectorsIDs.NOTIFICATIONS_BANNER),
-        ).toBeNull();
-      } finally {
-        jest.useRealTimers();
-      }
-    });
-  });
-
-  it('navigates to the socialAI notification settings section when notification button is pressed and preferences exist', () => {
-    renderWithProvider(<TopTradersView />);
-    fireEvent.press(
-      screen.getByTestId(TopTradersViewSelectorsIDs.NOTIFICATION_BUTTON),
-    );
-    expect(mockNavigate).toHaveBeenCalledWith(Routes.SETTINGS_VIEW, {
-      screen: Routes.SETTINGS.NOTIFICATION_SETTINGS_SECTION,
-      params: {
-        type: 'socialAI',
-        title: 'Trading signals',
-        description:
-          'Updates from traders and assets you follow, plus curated market news',
-      },
-    });
-  });
-
-  it('navigates to notification settings when preferences do not exist yet', () => {
-    mockHasNotificationPreferences.mockReturnValue(false);
-
-    renderWithProvider(<TopTradersView />);
-    fireEvent.press(
-      screen.getByTestId(TopTradersViewSelectorsIDs.NOTIFICATION_BUTTON),
-    );
-
-    expect(mockNavigate).toHaveBeenCalledWith(Routes.SETTINGS_VIEW, {
-      screen: Routes.SETTINGS.NOTIFICATIONS,
-    });
   });
 
   it('renders all traders', () => {
@@ -935,31 +739,6 @@ describe('TopTradersView', () => {
       screen.queryByTestId(TopTradersViewSelectorsIDs.TYPE_SELECTOR),
     ).toBeOnTheScreen();
     expect(screen.queryByText('alpha.eth')).not.toBeOnTheScreen();
-  });
-
-  describe('embedded in the Leaderboard | Feed tabs', () => {
-    it('scrolls the filters with the rows by placing them in the list header', () => {
-      renderWithProvider(<TopTradersView embeddedInTabs />);
-
-      expect(
-        screen.getByTestId(TopTradersViewSelectorsIDs.TYPE_SELECTOR),
-      ).toBeOnTheScreen();
-      expect(
-        screen.getByTestId(TopTradersViewSelectorsIDs.TRADER_LIST).props
-          .ListHeaderComponent,
-      ).not.toBeNull();
-    });
-
-    it('drops the large title and the scroll-pinned bar, which the tabs header replaces', () => {
-      renderWithProvider(<TopTradersView embeddedInTabs />);
-
-      expect(
-        screen.queryByTestId(TopTradersViewSelectorsIDs.TITLE),
-      ).not.toBeOnTheScreen();
-      expect(
-        screen.queryByTestId(TopTradersViewSelectorsIDs.PINNED_FILTER_BAR),
-      ).not.toBeOnTheScreen();
-    });
   });
 
   describe('performance', () => {
