+++ app/components/Views/SocialLeaderboard/index.ts
@@ -1,5 +1,4 @@
 export { default as TopTradersView } from './TopTradersView';
-export { default as SocialTradersView } from './SocialTradersView';
 export { default as SocialTradersTabsView } from './SocialTradersTabsView';
 export { default as FeedView } from './FeedView';
 export { default as TraderProfileView } from './TraderProfileView';
